Consolidate readme files and relocate project docs
There were redundant readmes with conflicting or out of date
information. This commit consolidates the README.md files to the one in
the root of the repo.

Moved the project docs (for e.g. contributing guidelines and other
community details) into the root of the repo with the intent that the
project can then build the usual docs websites. This does not relocate
or at all change the API docs that would be derived from the rust source
in libursa.

Signed-off-by: Dan Middleton <dan.middleton@intel.com>